# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Enjin Blockchain REST API",
    "description": "REST API for Enjin and Canary Relaychain and Matrixchain. Forked from Parity Technologies' Polkadot REST API.",
    "contact": {
      "url": "https://github.com/enjin/blockchain-rest-api"
    },
    "license": {
      "name": "GPL-3.0-or-later"
    },
    "version": "0.3.2"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Localhost"
    }
  ],
  "paths": {
...
    "/v1/blocks/head/header": {
      "get": {
        "tags": [
          "blocks"
        ],
        "summary": "Get head block header",
        "description": "Returns the header of the latest finalized or canonical block (lightweight, no extrinsics/events).",
        "operationId": "get_blocks_head_header",
        "parameters": [
          {
            "name": "finalized",
            "in": "query",
            "description": "When true (default), returns finalized head header. When false, returns canonical head header.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "useRcBlock",
            "in": "query",
            "description": "Treat as Relay Chain block and return Matrixchain blocks",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Block header information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Service unavailable"
          }
        }
      }
    },
    "/v1/blocks/{blockId}": {
      "get": {
        "tags": [
          "blocks"
        ],
        "summary": "Get block by ID",
        "description": "Returns block information for a given block identifier (hash or number), including extrinsics, events, and fees. An entry that could not be decoded is still returned at its own index, with `decodeError` set, no `method` or `args`, and `era` as an empty object, and the response carries `partial: true`; its `events`, `success` and `paysFee` come from the block's events, and `success` is false when the block carried no outcome event for that index. Events emitted after the last extrinsic, by the FRAME poll hook or a multi block migration, are returned in a separate `afterExtrinsics` object, omitted when the block has none.",
        "operationId": "get_block",
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "description": "Block height number or block hash",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventDocs",
            "in": "query",
            "description": "Include documentation for events",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "extrinsicDocs",
            "in": "query",
            "description": "Include documentation for extrinsics",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "noFees",
            "in": "query",
            "description": "Skip fee calculation for extrinsics",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "finalizedKey",
            "in": "query",
            "description": "When true (default), include finalized status in response",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "decodedXcmMsgs",
            "in": "query",
            "description": "Decode and include XCM messages",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "paraId",
            "in": "query",
            "description": "Filter XCM messages by parachain ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "useRcBlock",
            "in": "query",
            "description": "Treat blockId as Relay Chain block and return Matrixchain blocks",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "useEvmFormat",
            "in": "query",
            "description": "Convert AccountId32 addresses to EVM format for revive pallet events",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Block information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid block identifier"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Service unavailable"
          }
        }
      }
    },
...
    "/v1/rc/blocks/{blockId}": {
      "get": {
        "tags": [
          "rc"
        ],
        "summary": "RC get block by ID",
        "description": "Returns relay chain block information for a given block identifier. An entry that could not be decoded is still returned at its own index, with `decodeError` set, no `method` or `args`, and `era` as an empty object, and the response carries `partial: true`; its `events`, `success` and `paysFee` come from the block's events, and `success` is false when the block carried no outcome event for that index. Events emitted after the last extrinsic, by the FRAME poll hook or a multi block migration, are returned in a separate `afterExtrinsics` object, omitted when the block has none.",
        "operationId": "get_rc_block",
        "parameters": [
          {
            "name": "blockId",
            "in": "path",
            "description": "Block height number or block hash",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "eventDocs",
            "in": "query",
            "description": "Include event documentation",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "extrinsicDocs",
            "in": "query",
            "description": "Include extrinsic documentation",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "noFees",
            "in": "query",
            "description": "Skip fee calculation",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "decodedXcmMsgs",
            "in": "query",
            "description": "Include decoded XCM messages",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "paraId",
            "in": "query",
            "description": "Filter XCM messages by parachain ID",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Relay chain block information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Invalid block identifier"
          },
          "500": {
            "description": "Internal server error"
          },
          "503": {
            "description": "Relay chain not configured"
          }
        }
      }
    },
...
  },
...
}